+++ ChangeLog.md
@@ -26,6 +26,7 @@
 - Improve compile info output
 - Limit strings to column size
 - Make ImportService stable for multiple calls
+- Mark some options as deprecated
 - MySql: Allow configuration of engine, charset and collation
 - Quirks for Freebox Player
 - Provide bash completion file